feat: 现在食药计算器也可以使用 `查看报表` 功能
